updated schemas to use add_attr() on a schema object instead of adding schema fields to the data_type itself.
Fixed the sampling rate change test section to use a dataset instead of a data_entry.

Added the object type to the metadata constructor in the tests (the actual object still needs to be updated) such that Metadata(id, class,...)

Removed the updating of object id's after a unit of work is commited as the commit will update the id's itself.